+++ tests/test_poisson_enhanced.cpp
@@ -781,8 +781,8 @@ TEST_F(PoissonEnhancedTest, ParallelBatchFittingTests) {
     
     for (int t = 0; t < num_threads; ++t) {
         threads.emplace_back([&, t]() {
-            concurrent_results[t].resize(datasets.size());
-            PoissonDistribution::parallelBatchFit(datasets, concurrent_results[t]);
+            concurrent_results[static_cast<size_t>(t)].resize(datasets.size());
+            PoissonDistribution::parallelBatchFit(datasets, concurrent_results[static_cast<size_t>(t)]);
         });
     }
     
@@ -793,7 +793,7 @@ TEST_F(PoissonEnhancedTest, ParallelBatchFittingTests) {
     // Verify all concurrent results match
     for (int t = 0; t < num_threads; ++t) {
         for (size_t i = 0; i < datasets.size(); ++i) {
-            EXPECT_NEAR(concurrent_results[t][i].getLambda(), batch_results[i].getLambda(), 1e-10)
+            EXPECT_NEAR(concurrent_results[static_cast<size_t>(t)][i].getLambda(), batch_results[i].getLambda(), 1e-10)
                 << "Thread " << t << " lambda result mismatch for dataset " << i;
         }
     }
